Introduced the Aiolli prediction mechanism and used it to make predictions (which, unfortunately, aren't yet quite as good as the original naive predictions, but a lot of parameter tweaking to do). Made some changes to DataLoader

+++ src/uk/ac/cam/queens/w3/Product.java
@@ -11,7 +11,7 @@ import java.util.Comparator;
  */
 public class Product {
     private int productId;
-    private double count;
+    private int count;
     private double weightedCount;
 
     public Product(int productId) {
